LocationItem コンポーネントを独立ファイルに分割
